requirements.txt
fastapi==0.95.1
mysql-connector-python==8.0.33
